requirements.txt
discord.py==2.3.2
httpx==0.27.0
pytz==2024.1
audioop-lts==0.2.1
deep-translator==1.11.4
beautifulsoup4==4.12.3
